Optimise the Attributes, doesn't do the update if it is already done.

+++ src/main/java/io/github/minecraftgui/models/attributes/AttributeGroupDouble.java
@@ -42,6 +42,6 @@ public class AttributeGroupDouble extends AttributeGroup<Double> {
 
     @Override
     public Double getValue() {
-        return (Double) super.getValue();
+        return super.getValue();
     }
 }
